requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
SQLAlchemy==1.4.39
PyMySQL==1.0.2
python-dotenv==0.20.0
PyJWT==2.8.0
Werkzeug==2.2.2
bcrypt==4.0.1
python-multipart==0.0.6
Flask-SQLAlchemy==2.5.1
mysql-connector-python==8.0.33
